requirements.txt
requests==2.31.0
pyjwt[crypto]==2.8.0
python-dotenv==1.0.0
cryptography==41.0.7